p23 u1 ir u2 ir pavadinimu keitimas

+++ JAVA/pamoka22-java-izanga-kintamieji-duomenu-tipai-funkcijos/src/uzduotis3.java
@@ -2,20 +2,20 @@ import java.util.Arrays;
 
 public class uzduotis3 {
     public static void main(String[] args) {
-        double a1 = 17.35;
-        double b1 = 18;
-        double c1 = 20.47;
-        double d1 = 16.37;
+        float a1 = 17.35f;
+        float b1 = 18f;
+        float c1 = 20.47f;
+        float d1 = 16.37f;
         System.out.println("namoPerimetras = " + namoPerimetras(a1, b1, c1, d1));
-        double a2 = 2;
-        double b2 = 4;
-        double c2 = 2;
-        double d2 = 4.1;
+        float a2 = 2f;
+        float b2 = 4f;
+        float c2 = 2f;
+        float d2 = 4.1f;
         System.out.println("namoPerimetras = " + namoPerimetras(a2, b2, c2, d2));
     }
 
 
-    public static double namoPerimetras(double a, double b, double c, double d) {
+    public static float namoPerimetras(float a, float b, float c, float d) {
         return a + b + c + d;
     }
 }
